POC remove spawns compat in order to drop linkme
This allows using the crate in cbindgen without worrying about values
being dropped at link time.

This makes spawns-core a bit heavier but removes the complexity of
linkme and might solve #6
