--- spawns-compat/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "spawns-compat"
-version = "0.2.1"
-readme.workspace = true
-authors.workspace = true
-edition.workspace = true
-homepage.workspace = true
-license.workspace = true
-repository.workspace = true
-documentation.workspace = true
-description = "Async runtime detections for thread context task spawner"
-
-[features]
-default = []
-tokio = ["linkme", "spawns-core/compat", "dep:tokio"]
-smol = ["linkme", "spawns-core/compat", "dep:smol"]
-async-global-executor = ["linkme", "spawns-core/compat", "dep:async-global-executor"]
-
-[dependencies]
-linkme = { version = "0.3.25", optional = true }
-smol = { version = "2.0.0", optional = true }
-spawns-core = { path = "../spawns-core", version = "1.1.1", optional = true }
-tokio = { version = "1", features = ["rt"], optional = true }
-async-global-executor = { version =  "2", optional = true }
-
-[dev-dependencies]
-async-std = "1.12.0"
-futures-lite = "2.3.0"
-tokio = { version = "1.37.0", features = ["full"] }
-
-[package.metadata.docs.rs]
-all-features = true
