+++ spawns-core/Cargo.toml
@@ -12,14 +12,19 @@ description = "Async runtime agnostic thread context task spawner for Rust"
 
 [features]
 default = []
-compat = ["linkme"]
+compat = []
+tokio = ["dep:tokio"]
+smol = ["dep:smol"]
+async-global-executor = ["dep:async-global-executor"]
 panic-multiple-global-spawners = []
 test-compat-global1 = ["compat"]
 test-compat-global2 = ["compat", "test-compat-global1"]
 test-named-global = []
 
 [dependencies]
-linkme = { version = "0.3.25", optional = true }
+smol = { version = "2.0.0", optional = true }
+tokio = { version = "1", features = ["rt"], optional = true }
+async-global-executor = { version =  "2", optional = true }
 
 [dev-dependencies]
 futures = "0.3.30"
