+++ spawns-core/src/lib.rs
@@ -3,7 +3,5 @@ mod compat;
 mod spawn;
 mod task;
 
-#[cfg(feature = "compat")]
-pub use compat::*;
 pub use spawn::*;
 pub use task::*;
